CR-changes MDB-269: Code Review Mage_Catalog.

+++ app/code/core/Mage/Catalog/Model/Resource/Product/Flat.php
@@ -48,7 +48,7 @@ class Mage_Catalog_Model_Resource_Product_Flat extends Mage_Core_Model_Resource_
     protected function _construct()
     {
         $this->_init('catalog/product_flat', 'entity_id');
-        $this->_storeId = Mage::app()->getStore()->getId();
+        $this->_storeId = (int)Mage::app()->getStore()->getId();
     }
 
     /**
